early commit after first pass at brainstorming

+++ src/main/java/com/lambazon/service/ProductService.java
@@ -19,7 +19,7 @@ public class ProductService {
 		return productRepository.products();
 	}
 
-	public Product product(Integer id) {
+	public Product product(Long id) {
 		return productRepository.product(id);
 	}
 }
